[Modules] (Moved raters and models)
- Moved raters and models into appropriate modules.
