+++ test/integration/service/user/service_test.go
@@ -6,13 +6,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	repouser "genesis-currency-api/internal/repository/user"
+	"genesis-currency-api/internal/user/model"
+	repouser "genesis-currency-api/internal/user/repository"
 	servuser "genesis-currency-api/internal/user/service"
 	"log"
 	"path/filepath"
 	"testing"
 
-	"genesis-currency-api/internal/model"
 	"genesis-currency-api/pkg/config"
 	"genesis-currency-api/pkg/dto"
 	myerrors "genesis-currency-api/pkg/errors"
